turn integration into function
