--- lib/runIntegration.js
@@ -1,3 +1,0 @@
-const integration = require('../integration').integration;
-
-integration();
