+++ package.json
@@ -33,7 +33,7 @@
     "test-unit": "istanbul cover _mocha -- test/**",
     "test": "npm run lint && npm run test-unit",
     "fetch-data": "./fetch_data.sh",
-    "integrate-data": "DEBUG=integration node lib/runIntegration",
+    "integrate-data": "DEBUG=integration node integration",
     "prepare-data": "npm run fetch-data && npm run integrate-data",
     "postinstall": "npm run prepare-data"
   },
